requirements.txt
flask==1.1.1
numpy==1.16.2
opencv-python==4.2.0.32
pep257==0.7.0
Pillow==6.0.0
